requirements.txt
Pillow==2.7.0
ImageHash==0.3

